testing code climate communication (#2)
* add conventional method naming

* listing from one single page

* removed comments

* fixed NoMethod error

* update file names

* another attempt at scraping just one list of events

* correct method name

* listed the events with html tags

* fixed nomethod error

* remove html tags

* collecting just text

* separates the array elements with #{"e"} and a comma

* testing class variable scraper

* testing initialize @events

* fixing Argument error

* tried another scraper method

I need to stop and refactor

* updated README

* update version file

* updated git version

* fixed gem spec errors

* gem description added to gemspec

*  added ruby gem badge

* gemspec

* moved gem version badge

* cleaning up the MD

* MD corrections

* moved badge

* code climate badge

* removed badge

* added badge

* version update
