+++ Gemfile
@@ -2,4 +2,3 @@ source 'https://rubygems.org'
 
 # Specify your gem's dependencies in ntr_results.gemspec
 gemspec
-
