+++ lib/ntr_results.rb
@@ -5,6 +5,5 @@ require 'pry'
 require 'open-uri'
 
 require_relative "./ntr_results/version"
-require_relative './ntr_results/event'
+require_relative './ntr_results/scraper'
 require_relative './ntr_results/cli'
-# require_relative './ntr_results/scraper'
