+++ lib/ntr_results/version.rb
@@ -1,3 +1,3 @@
 module NtrResults
-  VERSION = "0.1.0"
+  VERSION = "2.0"
 end
