+++ ntr_results.gemspec
@@ -9,19 +9,18 @@ Gem::Specification.new do |spec|
   spec.authors       = ["thesilentc"]
   spec.email         = ["thesilentc@me.com"]
 
-  spec.summary       = %q{todo: Write a short summary, because Rubygems requires one.}
-  spec.description   = %q{todo: Write a longer description or delete this line.}
-  spec.homepage      = "todo: Put your gem's website or public repo URL here."
+  spec.summary       = %q{This Ruby Gem provides a CLI to view recent events from National Team Roping and gives the user the option of seeing the winners of those events.}
+  spec.homepage      = "https://github.com/thesilentc/ntr_results"
   spec.license       = "MIT"
 
   # Prevent pushing this gem to RubyGems.org. To allow pushes either set the 'allowed_push_host'
   # to allow pushing to a single host or delete this section to allow pushing to any host.
-    if spec.respond_to?(:metadata)
-      spec.metadata['allowed_push_host'] = "todo: Set to 'http://mygemserver.com'"
-    else
-      raise "RubyGems 2.0 or newer is required to protect against " \
-        "public gem pushes."
-    end
+    # if spec.respond_to?(:metadata)
+    #   spec.metadata['allowed_push_host'] = "todo: Set to 'http://mygemserver.com'"
+    # else
+    #   raise "RubyGems 2.0 or newer is required to protect against " \
+    #     "public gem pushes."
+    # end
 
     spec.files         = `git ls-files -z`.split("\x0").reject do |f|
       f.match(%r{^(test|spec|features)/})
